requirements.txt
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
ddgs>=1.0.0
pytest>=8.0.0
pyyaml>=6.0.0
requests>=2.31.0
# yt-dlp>=2024.1.0    # Agent 05 档 2 视频下载：当前未启用，需启用时取消注释
